requirements.txt
accelerate==1.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-tkinter-loop==0.9.3
attrs==24.3.0
Authlib==1.3.1
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.6
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.19.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.5
effdet==0.4.1
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.1.0
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.9.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-localai==0.1.2
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.2.10
layoutparser==0.3.4
llama-cloud==0.1.8
llama-index==0.12.10
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.10.post1
llama-index-embeddings-huggingface==0.5.0
llama-index-embeddings-ollama==0.5.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-groq==0.3.1
llama-index-llms-ollama==0.5.0
llama-index-llms-openai==0.3.13
llama-index-llms-openai-like==0.3.3
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
llamaparse==0.1.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
olefile==0.47
ollama==0.4.5
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.59.6
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pi_heif==0.21.0
pikepdf==9.5.1
pillow==11.1.0
pip==24.2
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.7.1
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.1
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-json-logger==3.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.0
semantic-chunkers==0.0.10
semantic-router==0.0.68
Send2Trash==1.8.3
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.13
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.21.0
toml==0.10.2
torch==2.7.0.dev20250114
torchaudio==2.6.0.dev20250114
torchvision==0.22.0.dev20250114
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
unstructured==0.16.13
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.10.2
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
